worker/requirements.txt
pika
netmiko==4.2.0
pymongo
python-dotenv
